Quick fix for #23208.
