+++ client/components/tinymce/plugins/media/plugin.jsx
@@ -27,7 +27,6 @@ import EditorMediaModal from 'post-editor/editor-media-modal';
 import notices from 'notices';
 import TinyMCEDropZone from './drop-zone';
 import restrictSize from './restrict-size';
-import { insertMediaAsLink, insertMedia } from './utils';
 import advanced from './advanced';
 import config from 'config';
 import { getSelectedSite } from 'state/ui/selectors';
@@ -58,6 +57,21 @@ function mediaButton( editor ) {
 		updateMedia; // eslint-disable-line
 
 	const getSelectedSiteFromState = () => getSelectedSite( getState() );
+
+	function insertMedia( markup ) {
+		editor.execCommand( 'mceInsertContent', false, markup );
+	}
+
+	function insertMediaAsLink( media ) {
+		const mediaUrl = media.URL;
+		const title = media.title.replace( /"/g, '&quot;' );
+		editor.execCommand(
+			'mceReplaceContent',
+			false,
+			`<a href="${ mediaUrl }"
+			title="${ title }">{$selection}</a>`
+		);
+	}
 
 	function renderModal( props = {}, options = {} ) {
 		const selectedSite = getSelectedSiteFromState();
@@ -94,9 +108,9 @@ function mediaButton( editor ) {
 				onInsertMedia={ ( markup, media ) => {
 					const selectedText = editor.selection.getContent( { format: 'text' } );
 					if ( media.length === 1 && selectedText ) {
-						insertMediaAsLink( editor, media[ 0 ] );
+						insertMediaAsLink( media[ 0 ] );
 					} else {
-						insertMedia( editor, markup );
+						insertMedia( markup );
 					}
 					renderModal( { visible: false } );
 				} }
